feat!: Remove logs rolling. (#4191)
* feat!: Remove logs rolling.

+++ packages/runtime/lib/runtime.js
@@ -4,6 +4,7 @@ const { ITC } = require('@platformatic/itc')
 const {
   features,
   ensureLoggableError,
+  ensureError,
   executeWithTimeout,
   deepmerge,
   parseMemorySize,
@@ -11,14 +12,13 @@ const {
   kMetadata
 } = require('@platformatic/foundation')
 const { once, EventEmitter } = require('node:events')
-const { createReadStream, watch, existsSync } = require('node:fs')
-const { readdir, readFile, stat, access } = require('node:fs/promises')
+const { existsSync } = require('node:fs')
+const { readFile } = require('node:fs/promises')
 const { STATUS_CODES } = require('node:http')
 const { join } = require('node:path')
 const { pathToFileURL } = require('node:url')
 const { setTimeout: sleep, setImmediate: immediate } = require('node:timers/promises')
 const { Worker } = require('node:worker_threads')
-const ts = require('tail-file-stream')
 const { Agent, interceptors: undiciInterceptors, request } = require('undici')
 const { createThreadInterceptor } = require('undici-thread-interceptor')
 const SonicBoom = require('sonic-boom')
@@ -29,7 +29,7 @@ const { startManagementApi } = require('./management-api')
 const { startPrometheusServer } = require('./prom-server')
 const { startScheduler } = require('./scheduler')
 const { createSharedStore } = require('./shared-http-cache')
-const { getRuntimeTmpDir, getRuntimeLogsDir } = require('./utils')
+const { getRuntimeTmpDir } = require('./utils')
 const { sendViaITC, waitEventFromITC } = require('./worker/itc')
 const { RoundRobinMap } = require('./worker/round-robin-map.js')
 const {
@@ -71,7 +71,6 @@ class Runtime extends EventEmitter {
   #context
   #isProduction
   #runtimeTmpDir
-  #runtimeLogsDir
   #servicesIds
   #entrypointId
   #url
@@ -104,7 +103,6 @@ class Runtime extends EventEmitter {
     this.#context = context ?? {}
     this.#isProduction = this.#context.isProduction ?? this.#context.production ?? false
     this.#runtimeTmpDir = getRuntimeTmpDir(this.#root)
-    this.#runtimeLogsDir = getRuntimeLogsDir(this.#root, process.pid)
     this.#workers = new RoundRobinMap()
     this.#servicesIds = []
     this.#url = undefined
@@ -161,7 +159,7 @@ class Runtime extends EventEmitter {
     }
 
     // Create the logger
-    const [logger, destination] = await createLogger(config, this.#runtimeLogsDir)
+    const [logger, destination] = await createLogger(config)
     this.logger = logger
     this.#loggerDestination = destination
 
@@ -383,12 +381,8 @@ class Runtime extends EventEmitter {
 
     await this.stop(silent)
 
-    if (this.#managementApi) {
-      // This allow a close request coming from the management API to correctly be handled
-      setImmediate(() => {
-        this.#managementApi.close()
-      })
-    }
+    // The management API autocloses by itself via event in management-api.js.
+    // This is needed to let management API stop endpoint to reply.
 
     if (this.#prometheusServer) {
       await this.#prometheusServer.close()
@@ -553,8 +547,7 @@ class Runtime extends EventEmitter {
         metrics = await this.getFormattedMetrics()
       } catch (error) {
         if (!(error instanceof errors.RuntimeExitedError)) {
-          // TODO(mcollina): use the logger
-          console.error('Error collecting metrics', error)
+          this.logger.error({ err: ensureLoggableError(error) }, 'Error collecting metrics')
         }
         return
       }
@@ -567,87 +560,18 @@ class Runtime extends EventEmitter {
     }, COLLECT_METRICS_TIMEOUT).unref()
   }
 
-  async pipeLogsStream (writableStream, logger, startLogId, endLogId, runtimePID) {
-    endLogId = endLogId || Infinity
-    runtimePID = runtimePID ?? process.pid
+  async addLoggerDestination (writableStream) {
+    // Add the stream - We output everything we get
+    this.#loggerDestination.add({ stream: writableStream, level: 1 })
 
-    const runtimeLogFiles = await this.#getRuntimeLogFiles(runtimePID)
-
-    if (runtimeLogFiles.length === 0) {
-      writableStream.end()
-      return
-    }
-
-    let latestFileId = parseInt(runtimeLogFiles.at(-1).slice('logs.'.length))
-
-    let fileStream = null
-    let fileId = startLogId ?? latestFileId
-    let isClosed = false
-
-    const runtimeLogsDir = this.#getRuntimeLogsDir(runtimePID)
-
-    const watcher = watch(runtimeLogsDir, async (event, filename) => {
-      if (event === 'rename' && filename.startsWith('logs')) {
-        const logFileId = parseInt(filename.slice('logs.'.length))
-        if (logFileId > latestFileId) {
-          latestFileId = logFileId
-          fileStream.unwatch()
-        }
-      }
-    }).unref()
-
-    const streamLogFile = () => {
-      if (fileId > endLogId) {
-        writableStream.end()
-        return
-      }
-
-      const fileName = 'logs.' + fileId
-      const filePath = join(runtimeLogsDir, fileName)
-
-      const prevFileStream = fileStream
-
-      fileStream = ts.createReadStream(filePath)
-      fileStream.pipe(writableStream, { end: false, persistent: false })
-
-      if (prevFileStream) {
-        prevFileStream.unpipe(writableStream)
-        prevFileStream.destroy()
-      }
-
-      fileStream.on('close', () => {
-        if (latestFileId > fileId && !isClosed) {
-          streamLogFile(++fileId)
-        }
-      })
-
-      fileStream.on('error', err => {
-        isClosed = true
-        logger.error(err, 'Error streaming log file')
-        fileStream.destroy()
-        watcher.close()
-        writableStream.end()
-      })
-
-      fileStream.on('eof', () => {
-        if (fileId >= endLogId) {
-          writableStream.end()
-          return
-        }
-        if (latestFileId > fileId) {
-          fileStream.unwatch()
-        }
-      })
-
-      return fileStream
-    }
-
-    streamLogFile(fileId)
+    // Immediately get the counter of the lastId so we can use it to later remove it
+    const id = this.#loggerDestination.lastId
 
     const onClose = () => {
-      isClosed = true
-      watcher.close()
-      fileStream.destroy()
+      writableStream.removeListener('close', onClose)
+      writableStream.removeListener('error', onClose)
+      this.removeListener('closed', onClose)
+      this.#loggerDestination.remove(id)
     }
 
     writableStream.on('close', onClose)
@@ -1008,44 +932,6 @@ class Runtime extends EventEmitter {
     this.servicesConfigsPatches.delete(id)
   }
 
-  async getLogIds (runtimePID) {
-    runtimePID = runtimePID ?? process.pid
-
-    const runtimeLogFiles = await this.#getRuntimeLogFiles(runtimePID)
-    const runtimeLogIds = []
-
-    for (const logFile of runtimeLogFiles) {
-      const logId = parseInt(logFile.slice('logs.'.length))
-      runtimeLogIds.push(logId)
-    }
-    return runtimeLogIds
-  }
-
-  async getAllLogIds () {
-    const runtimesLogFiles = await this.#getAllLogsFiles()
-    const runtimesLogsIds = []
-
-    for (const runtime of runtimesLogFiles) {
-      const runtimeLogIds = []
-      for (const logFile of runtime.runtimeLogFiles) {
-        const logId = parseInt(logFile.slice('logs.'.length))
-        runtimeLogIds.push(logId)
-      }
-      runtimesLogsIds.push({
-        pid: runtime.runtimePID,
-        indexes: runtimeLogIds
-      })
-    }
-
-    return runtimesLogsIds
-  }
-
-  async getLogFileStream (logFileId, runtimePID) {
-    const runtimeLogsDir = this.#getRuntimeLogsDir(runtimePID)
-    const filePath = join(runtimeLogsDir, `logs.${logFileId}`)
-    return createReadStream(filePath)
-  }
-
   #getHttpCacheValue ({ request }) {
     if (!this.#sharedHttpCache) {
       return
@@ -1257,8 +1143,7 @@ class Runtime extends EventEmitter {
           count: workersCount
         },
         inspectorOptions,
-        dirname: this.#root,
-        runtimeLogsDir: this.#runtimeLogsDir
+        dirname: this.#root
       },
       argv: serviceConfig.arguments,
       execArgv,
@@ -1580,7 +1465,9 @@ class Runtime extends EventEmitter {
           gracePeriod > 0 ? gracePeriod : 1
         )
       }
-    } catch (error) {
+    } catch (err) {
+      const error = ensureError(err)
+
       // TODO: handle port allocation error here
       if (error.code === 'EADDRINUSE' || error.code === 'EACCES') throw error
 
@@ -1589,7 +1476,13 @@ class Runtime extends EventEmitter {
       if (worker[kWorkerStatus] !== 'exited') {
         // This prevent the exit handler to restart service
         worker[kWorkerStatus] = 'exited'
-        await worker.terminate()
+
+        // Wait for the worker to exit gracefully, otherwise we terminate it
+        const waitTimeout = await executeWithTimeout(once(worker, 'exit'), config.gracefulShutdown.service)
+
+        if (waitTimeout === kTimeout) {
+          await worker.terminate()
+        }
       }
 
       this.emit('service:worker:start:error', { ...eventPayload, error })
@@ -1900,49 +1793,6 @@ class Runtime extends EventEmitter {
     const packageJsonFile = await readFile(packageJsonPath, 'utf8')
     const packageJson = JSON.parse(packageJsonFile)
     return packageJson
-  }
-
-  #getRuntimeLogsDir (runtimePID) {
-    return join(this.#runtimeTmpDir, runtimePID.toString(), 'logs')
-  }
-
-  async #getRuntimeLogFiles (runtimePID) {
-    const runtimeLogsDir = this.#getRuntimeLogsDir(runtimePID)
-    const runtimeLogsFiles = await readdir(runtimeLogsDir)
-    return runtimeLogsFiles
-      .filter(file => file.startsWith('logs'))
-      .sort((log1, log2) => {
-        const index1 = parseInt(log1.slice('logs.'.length))
-        const index2 = parseInt(log2.slice('logs.'.length))
-        return index1 - index2
-      })
-  }
-
-  async #getAllLogsFiles () {
-    try {
-      await access(this.#runtimeTmpDir)
-    } catch (err) {
-      this.logger.error({ err: ensureLoggableError(err) }, 'Cannot access temporary folder.')
-      return []
-    }
-
-    const runtimePIDs = await readdir(this.#runtimeTmpDir)
-    const runtimesLogFiles = []
-
-    for (const runtimePID of runtimePIDs) {
-      const runtimeLogsDir = this.#getRuntimeLogsDir(runtimePID)
-      const runtimeLogsDirStat = await stat(runtimeLogsDir)
-      const runtimeLogFiles = await this.#getRuntimeLogFiles(runtimePID)
-      const lastModified = runtimeLogsDirStat.mtime
-
-      runtimesLogFiles.push({
-        runtimePID: parseInt(runtimePID),
-        runtimeLogFiles,
-        lastModified
-      })
-    }
-
-    return runtimesLogFiles.sort((runtime1, runtime2) => runtime1.lastModified - runtime2.lastModified)
   }
 
   #handleWorkerStandardStreams (worker, serviceId, workerId) {
